Break out the kernel manager and messages
