+++ test/src/kernel/kernel.spec.ts
@@ -1806,171 +1806,6 @@ describe('kernel', () => {
     });
   });
 
-  describe('KernelManager', () => {
-
-    describe('#constructor()', () => {
-
-      it('should take the options as an argument', () => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        expect(manager instanceof KernelManager).to.be(true);
-      });
-
-    });
-
-    describe('#specsChanged', () => {
-
-      it('should be emitted when the specs change', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        manager.specsChanged.connect((sender, args) => {
-          expect(sender).to.be(manager);
-          expect(deepEqual(args, KERNELSPECS)).to.be(true);
-          done();
-        });
-        let handler = new RequestHandler(() => {
-          handler.respond(200, KERNELSPECS);
-        });
-        manager.getSpecs();
-      });
-
-    });
-
-    describe('#runningChanged', () => {
-
-      it('should be emitted in listRunning when the running kernels changed', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let data: Kernel.IModel[] = [
-          { id: uuid(), name: 'test' },
-          { id: uuid(), name: 'test2' }
-        ];
-        manager.runningChanged.connect((sender, args) => {
-          expect(sender).to.be(manager);
-          expect(deepEqual(args, data)).to.be(true);
-          done();
-        });
-        let handler = new RequestHandler(() => {
-          handler.respond(200, data);
-        });
-        manager.listRunning();
-      });
-
-    });
-
-    describe('#getSpecs()', () => {
-
-      it('should get the list of kernel specs', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let ids = {
-          'python': PYTHON_SPEC,
-          'python3': PYTHON3_SPEC
-        };
-        let handler = new RequestHandler(() => {
-          handler.respond(200, { 'default': 'python',
-                               'kernelspecs': ids });
-        });
-        manager.getSpecs().then(specs => {
-          let names = Object.keys(specs.kernelspecs);
-          expect(names[0]).to.be('python');
-          expect(names[1]).to.be('python3');
-          done();
-        });
-      });
-
-    });
-
-    describe('#listRunning()', () => {
-
-      it('should list the running kernels', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let data = [
-          { id: uuid(), name: 'test' },
-          { id: uuid(), name: 'test2' }
-        ];
-        let handler = new RequestHandler(() => {
-          handler.respond(200, data);
-        });
-        manager.listRunning().then(response => {
-          expect(response[0]).to.eql(data[0]);
-          expect(response[1]).to.eql(data[1]);
-          done();
-        });
-      });
-
-    });
-
-    describe('#startNew()', () => {
-
-      it('should start a new kernel', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let tester = new KernelTester(() => {
-          tester.respond(201, { id: uuid(), name: KERNEL_OPTIONS.name });
-        });
-        manager.startNew().then(kernel => {
-          expect(kernel.status).to.be('unknown');
-          kernel.statusChanged.connect(() => {
-            if (kernel.status === 'starting') {
-              kernel.dispose();
-              done();
-            }
-          });
-        });
-
-      });
-
-    });
-
-    describe('#findById()', () => {
-
-      it('should find an existing kernel by id', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let id = uuid();
-        let tester = new KernelTester(() => {
-          tester.respond(201, { id: id, name: KERNEL_OPTIONS.name });
-        });
-        manager.startNew().then(kernel => {
-          manager.findById(id).then(newKernel => {
-            expect(newKernel.name).to.be(kernel.name);
-            expect(newKernel.id).to.be(kernel.id);
-            done();
-          });
-        });
-      });
-
-    });
-
-    describe('#connectTo()', () => {
-
-      it('should connect to an existing kernel', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let id = uuid();
-        let tester = new KernelTester(() => {
-          tester.respond(201, { id: id, name: KERNEL_OPTIONS.name });
-        });
-        manager.startNew().then(kernel => {
-          manager.connectTo(id).then(newKernel => {
-            expect(newKernel.name).to.be(kernel.name);
-            expect(newKernel.id).to.be(kernel.id);
-            expect(newKernel).to.not.be(kernel);
-            done();
-          });
-        });
-      });
-
-    });
-
-    describe('shutdown()', () => {
-
-      it('should shut down a kernel by id', (done) => {
-        let manager = new KernelManager(KERNEL_OPTIONS);
-        let handler = new RequestHandler(() => {
-          handler.respond(204, { });
-        });
-        manager.shutdown('foo').then(done, done);
-      });
-
-    });
-
-  });
-
   describe('Kernel.getSpecs()', () => {
 
     it('should load the kernelspecs', (done) => {
@@ -2101,141 +1936,6 @@ describe('kernel', () => {
       });
       let promise = Kernel.getSpecs();
       expectAjaxError(promise, done, 'Invalid Status: 201');
-    });
-
-  });
-
-  describe('#isStreamMsg()', () => {
-
-    it('should check for a stream message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'stream', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isStreamMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isStreamMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isDisplayDataMsg()', () => {
-
-    it('should check for a display data message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'display_data', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isDisplayDataMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isDisplayDataMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isExecuteInputMsg()', () => {
-
-    it('should check for a execute input message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'execute_input', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isExecuteInputMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isExecuteInputMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isExecuteResultMsg()', () => {
-
-    it('should check for an execute result message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'execute_result', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isExecuteResultMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isExecuteResultMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isStatusMsg()', () => {
-
-    it('should check for a status message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'status', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isStatusMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isStatusMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isClearOutputMsg()', () => {
-
-    it('should check for a clear output message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'clear_output', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isClearOutputMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isClearOutputMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isCommOpenMsg()', () => {
-
-    it('should check for a comm open message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'comm_open', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isCommOpenMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isCommOpenMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isErrorMsg()', () => {
-
-    it('should check for an message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'error', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isErrorMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'iopub', session: 'baz'
-      });
-      expect(KernelMessage.isErrorMsg(msg)).to.be(false);
-    });
-
-  });
-
-  describe('#isInputRequestMsg()', () => {
-
-    it('should check for an input_request message type', () => {
-      let msg = KernelMessage.createMessage({
-        msgType: 'input_request', channel: 'stdin', session: 'baz'
-      });
-      expect(KernelMessage.isInputRequestMsg(msg)).to.be(true);
-      msg = KernelMessage.createMessage({
-        msgType: 'foo', channel: 'stdin', session: 'baz'
-      });
-      expect(KernelMessage.isStatusMsg(msg)).to.be(false);
     });
 
   });
